Cp1/src/main/java/cvut/controllers/ReviewController.java
package cvut.controllers;

import cvut.model.Review;
import cvut.services.ReviewService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.lang.NonNull;
import org.springframework.security.access.prepost.PostFilter;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.List;

@RestController
@RequestMapping(path = "critiques")
@Validated
@RequiredArgsConstructor
@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
public class ReviewController {
    private final ReviewService reviewService;
    private static final Logger LOG = LoggerFactory.getLogger(ReviewController.class);
    @GetMapping(value = "/{reviewId}", produces = MediaType.APPLICATION_JSON_VALUE)
    @PreAuthorize("(!#user.isAdmin() && anonymous) || hasRole('ROLE_ADMIN')")
    public ResponseEntity<Review> getReviewById(@PathVariable("reviewId") @NonNull String reviewId) {
        Review review = reviewService.findById(Integer.parseInt(reviewId));
        LOG.debug("getReviewById: {}", review);

        if (review == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
        return new ResponseEntity<>(review, HttpStatus.OK);
    }

    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
    @PostFilter("hasRole('ROLE_ADMIN')")
    public List<Review> fetchAll() {
        return reviewService.findAll();
    }
}